Move util fn to tests
I removed the last real use a few commits ago, but it's still needed for these
tests, so I'm moving it here for future easier deletion
